Restructured components into /dash and /stash folders
