+++ app/javascript/react/components/Sidebar.js
@@ -1,15 +1,15 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import { Link } from 'react-router-dom'
 
-import NewWidgetForm from './NewWidgetForm'
+import NewWidgetForm from './dashboard/NewWidgetForm'
 
-const Sidebar = ({ activePath, addWidget, toggleEditMode, editMode, errors }) => {
+const Sidebar = ({ rootPath, addWidget, toggleEditMode, editMode, errors }) => {
   let dashClass = "side-bar-icon"
   let stashClass = "side-bar-icon"
 
   let addAWidgetButton, toggleEditButton
 
-  if(activePath === "/dash") {
+  if(rootPath === "/dash") {
     dashClass+= " active-icon"
     addAWidgetButton = <li className="side-bar-icon">
         <NewWidgetForm addWidget={addWidget} errors={errors} />
@@ -27,24 +27,26 @@ const Sidebar = ({ activePath, addWidget, toggleEditMode, editMode, errors }) =>
       </li>
     }
   }
-  if(activePath === "/stash") stashClass+= " active-icon"
+  if(rootPath === "/stash") stashClass+= " active-icon"
   return (
-    <nav id="side-bar">
-      <Link to="/dash">
-        <li className={dashClass}>
-          <i className="fas fa-chalkboard"></i>
-          <div className="iconTextHover">Dash</div>
-        </li>
-      </Link>
-      <Link to="/stash">
-        <li className={stashClass}>
-          <i className="fas fa-briefcase"></i>
-          <div className="iconTextHover">Stash</div>
-        </li>
-      </Link>
-      {addAWidgetButton}
-      {toggleEditButton}
-    </nav>
+    <Fragment>
+      <nav id="side-bar">
+        <Link to="/dash">
+          <li className={dashClass}>
+            <i className="fas fa-chalkboard"></i>
+            <div className="iconTextHover">Dash</div>
+          </li>
+        </Link>
+        <Link to="/stash">
+          <li className={stashClass}>
+            <i className="fas fa-briefcase"></i>
+            <div className="iconTextHover">Stash</div>
+          </li>
+        </Link>
+        {addAWidgetButton}
+        {toggleEditButton}
+      </nav>
+    </Fragment>
   )
 }
