+++ app/javascript/react/components/Topbar.js
@@ -1,118 +1,18 @@
-import React, { Fragment, useState, useEffect, useCallback } from 'react';
+import React, { Fragment, useState, useEffect } from 'react';
 import { Route, Switch, Link, Redirect } from 'react-router-dom';
 
-import Dashboard from './Dashboard'
-import Stash from './Stash'
-import JournalsIndexContainer from './journal/JournalsIndexContainer'
-import JournalShowContainer from './journal/JournalShowContainer'
-import Sidebar from './Sidebar'
+import DashboardContainer from './dashboard/DashboardContainer'
+import StashContainer from './stash/StashContainer'
 import ModalForm from './ModalForm'
 import LandingPageContainer from './LandingPageContainer'
+import LoginForm from './LoginForm'
+import JournalsIndexContainer from './stash/journal/JournalsIndexContainer'
+import JournalShowContainer from './stash/journal/JournalShowContainer'
 
 const Topbar = props => {
   const [currentUser, setCurrentUser] = useState(null)
   const [shouldLogout, setShouldLogout] = useState(false)
-  const [widgets, setWidgets] = useState([])
-  const [widgetErrors, setWidgetErrors] = useState([])
-  const [editMode, setEditMode] = useState(false)
-  const toggleEditMode = () => {
-    setEditMode(!editMode)
-  }
-
-  const moveWidget = useCallback(
-      (dragIndex, hoverIndex) => {
-        const dragWidget = widgets[dragIndex]
-        const newOrder = widgets.map((widget, index, oldOrder) => {
-          if (index === hoverIndex) {
-            return oldOrder[dragIndex]
-          }
-          if (index === dragIndex) {
-            return oldOrder[hoverIndex]
-          }
-          return widget
-        })
-        const liveUpdate = () => {
-          setWidgets(newOrder)
-        }
-        updateOrder({dragIndex: dragIndex, hoverIndex: hoverIndex}, liveUpdate)
-      },
-      [widgets],
-  )
-
-  const updateOrder = (payload, liveUpdate) => {
-    fetch('/api/v1/widgets/reorder', {
-      credentials: 'same-origin',
-      method: "PATCH",
-      headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json'
-      },
-      body: JSON.stringify(payload)
-    })
-    .then(response => {
-      if (response.ok) {
-        return response.json()
-      } else {
-        throw new Error(`${response.status} ${response.statusText}`)
-      }
-    })
-    .then(parsedBody => {
-      if (!Array.isArray(parsedBody)) {
-        liveUpdate()
-      }
-    })
-    .catch(error => console.error(`Error in widget order patch ${error.message}`))
-  }
   const activePath = props.location.pathname;
-
-  const addWidget = (payload, clearForm) => {
-    payload.position = widgets.length
-    fetch('/api/v1/widgets', {
-      credentials: 'same-origin',
-      method: "POST",
-      headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json'
-      },
-      body: JSON.stringify(payload)
-    })
-    .then(response => {
-      if (response.ok) {
-        return response.json();
-      } else {
-        throw new Error(`${response.status} ${response.statusText}`);
-      }
-    })
-    .then(parsedBody => {
-      if (!Array.isArray(parsedBody)) {
-        $('#widgetModal').modal('hide')
-        if (clearForm) clearForm()
-        setWidgetErrors([])
-        setWidgets([...widgets, parsedBody])
-      } else {
-        setWidgetErrors(parsedBody)
-      }
-    })
-    .catch(error => console.error(`Error in journal post fetch ${error.message}`))
-  }
-
-  const deleteWidget = (widgetId) => {
-    fetch(`/api/v1/widgets/${widgetId}`, {
-      credentials: 'same-origin',
-      method: "DELETE"
-    })
-    .then(response => {
-      if (response.ok) {
-        return response.json();
-      } else {
-        throw new Error(`${response.status} ${response.statusText}`);
-      }
-    })
-    .then(parsedBody => {
-      setWidgets(widgets.filter(widget => widget.id !== parsedBody.id))
-    })
-    .catch(error => console.error(`Error in widget delete fetch ${error.message}`))
-  }
 
   useEffect(() => {
     fetch('/api/v1/users/current')
@@ -125,19 +25,6 @@ const Topbar = props => {
     })
     .then(parsedBody => {
       setCurrentUser(parsedBody)
-    })
-    .catch(error => `Error in fetch ${error.message}`)
-
-    fetch('/api/v1/widgets')
-    .then(response => {
-      if (response.ok) {
-        return response.json()
-      } else {
-        throw new Error(`${response.status} ${response.statusText}`)
-      }
-    })
-    .then(parsedBody => {
-      setWidgets(parsedBody)
     })
     .catch(error => `Error in fetch ${error.message}`)
   }, [])
@@ -160,7 +47,6 @@ const Topbar = props => {
   if (shouldLogout && props.history.action !== "REPLACE") {
     window.location.replace("/welcome")
   }
-
   let togglerIcon =
   <button className="navbar-toggler" type="button" data-toggle="collapse" data-target=".dual-collapse2">
     <span className="navbar-toggler-icon"></span>
@@ -169,9 +55,9 @@ const Topbar = props => {
   let stashClass = "nav-item navbar-text navbar-sidebar"
   if(activePath === "/dash") dashClass+= " active"
   if(activePath === "/stash") stashClass+= " active"
-  let rightTopbarContent, leftTopbarContent, sidebar
+  let rightTopbarContent, leftTopbarContent, brandLink
   if(currentUser) {
-    sidebar = <Sidebar activePath={activePath} addWidget={addWidget} toggleEditMode={toggleEditMode}  editMode={editMode} errors={widgetErrors} />
+    brandLink = "/dash"
     togglerIcon =
     <button className="navbar-toggler" style={{padding:0, border:'none'}} type="button" data-toggle="collapse" data-target=".dual-collapse2">
       <img className="top-bar-profile-photo" src={currentUser.profilePhoto} />
@@ -208,7 +94,7 @@ const Topbar = props => {
               </ul>
           </div>
           <div className="mx-auto order-0">
-              <Link to="#" className="navbar-brand mx-auto" href="#">Dashy</Link>
+              <Link to={brandLink} className="navbar-brand mx-auto" href="#">Dashy</Link>
           </div>
           {togglerIcon}
           <div className="navbar-collapse collapse w-100 order-3">
@@ -218,12 +104,12 @@ const Topbar = props => {
           </div>
       </nav>
       <section className="display-area">
-        {sidebar}
         <Switch>
           <Route exact path='/' render={(props) => <LandingPageContainer user={currentUser} /> }/>
           <Route exact path='/welcome' render={(props) => <LandingPageContainer user={currentUser} />}/>
-          <Route exact path='/dash' render={(props) => <Dashboard widgets={widgets} moveWidget={moveWidget} deleteWidget={deleteWidget} editMode={editMode} /> }/>
-          <Route exact path='/stash' component={Stash}/>
+          <Route exact path='/dash' render={(props) => <DashboardContainer /> }/>
+          <Route exact path='/users/login' render={(props) => <LandingPageContainer user={currentUser} />}   />
+          <Route exact path='/stash' component={StashContainer} />
           <Route exact path='/stash/journals' component={JournalsIndexContainer}/>
           <Route exact path='/stash/journals/:id' component={JournalShowContainer}/>
         </Switch>
