+++ app/javascript/react/components/dashboard/widgets/JournalWidget.js
