+++ app/javascript/react/components/dashboard/widgets/RSSWidget.js
