+++ app/javascript/react/components/dashboard/widgets/WeatherWidget.js
