+++ app/javascript/react/components/stash/StashTile.js
