+++ app/javascript/react/components/stash/StashTile.test.js
