+++ app/javascript/react/components/stash/journal/JournalDetails.js
