+++ app/javascript/react/components/stash/journal/JournalNewForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 
-import ErrorList from '../ErrorList';
+import ErrorList from '../../ErrorList';
 
 const JournalNewForm = ({ handleFormDisplay, errors, addNewJournal, editJournal, oldJournal }) => {
   let defaultJournal = {
