+++ app/javascript/react/components/stash/journal/JournalShowContainer.js
