+++ app/javascript/react/components/stash/journal/JournalTile.js
