+++ app/javascript/react/components/stash/journal/JournalTile.test.js
